Convert one Karma "mock" test to use spies instead
This basically follows the model in the Angular testing documentation (https://angular.io/guide/testing-services#testing-http-services). I haven't really thought about it very hard so I don't have a clear sense of whether I prefer this approach, or why I might, but if nothing else it better aligns with the documentation students are likely to find.