Improve new `getUsers()` test
This converts the test of `getUsers()` with no parameters to the version suggested by @helloworld12321 (https://github.com/UMM-CSci-3601/3601-lab3_angular-javalin/pull/759#pullrequestreview-880686742). I had to fix one problem (the expected arguments to the call), and I added some comments to try to clarify what's going on.

I really like this change, and will probably try to apply it to the other tests in this file as well.