Convert remaining `UserService#get` tests to the new spy approach
This converts the other two `get()` tests to the new spy approach. The latter sometimes fails, though, because the order of params shouldn't matter, but the `.toHaveBeenCalledWith()` call totally cares about order.

I think the solution is to somehow capture what param map was actually passed as an argument, and then have expectations on that, but I'm not sure how to do that.