Reformat, reorganize, and improve explanations of tests
There is a lot going on here! This commit focuses on giving a bit more detail in the descriptions of what each group of tests is checking. It also realigns some of the blocks that got a little messed up so that you can collapse just some parts and still get the gist of what is happening, which can be helpful. Having an overview is really handy sometimes! Finally, it removes asyncData, which went away when we used of() as suggested by @helloworld12321 in the discussion of the PR.