Fix problems caused by Chai assert being seen before Jasmine assert
There were lots of problems showing up as red, complaining, and highlighting in all kinds of ways even though the component tests worked just fine. This was a known problem, and a suggested workaround was to put an `exclude` in `tsconfig.json` that makes it so that Chai assertions that are supplied so helpfully by Cypress won't accidentally make it so that Jasmine assertions are hidden.